Version Beta 0.3.2! / Вєрсия Beta 0.3.2! - replaced ulong for money with BigInteger; -- adapted the converter and methods for it; - the "Pow.cs" class represents large numbers (millions, trillions ...); - refactoring the Converter.cs class; -- more versatile and aesthetic methods; ---------------------------------------- - заменил ulong для денег на BigInteger; -- адаптировал конвертер и методы под него; - класс "Pow.cs" представляет большие числа (миллионы, триллионы ...); - рефакторинг класса Converter.cs; -- более универсальные и эстетичные методы;

+++ Assets/Scripts/Stats/StatsSystem.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Numerics;
 using UnityEngine;
 using UnityEngine.Events;
 
@@ -13,7 +14,7 @@ namespace PlanetsColony
 
         public static StatsSystem Instance = null;
 
-        [SerializeField] private ulong _money = 0;
+        [SerializeField] private BigInteger _money = 0;
         [SerializeField] private uint _maxShipsCount = 10;
         [SerializeField] private uint _activeShipsCount = 0;
 
@@ -31,7 +32,7 @@ namespace PlanetsColony
             OnMoneyValueChange?.Invoke();
         }
 
-        public ulong GetMoney()
+        public BigInteger GetMoney()
         {
             return _money;
         }
